update and split types in new files
